+++ lib/const/const.dart
@@ -3,44 +3,52 @@ import 'package:flutter/material.dart';
 const String refreshEvent = 'refreshEvent';
 
 /// icons
-class ConstIcon {
+class WayIcons {
+  WayIcons._();
+
   /// 箭头右
-  static const IconData arrowRight = _IconCode(0xe65b);
+  static const IconData arrowRight = _IconData(0xe657);
 
   /// 箭头左
-  static const IconData arrowLeft = _IconCode(0xe659);
+  static const IconData arrowLeft = _IconData(0xe656);
 
   /// 箭头上
-  static const IconData arrowUp = _IconCode(0xe658);
+  static const IconData arrowUp = _IconData(0xe658);
 
   /// 箭头下
-  static const IconData arrowDown = _IconCode(0xe65a);
+  static const IconData arrowDown = _IconData(0xe655);
 
   /// 搜索
-  static const IconData search = _IconCode(0xe8ba);
+  static const IconData search = _IconData(0xe65f);
 
   /// 成功
-  static const IconData success = _IconCode(0xe645);
+  static const IconData success = _IconData(0xe660);
 
   /// 错误
-  static const IconData fail = _IconCode(0xe669);
+  static const IconData fail = _IconData(0xe65d);
 
   /// 提示
-  static const IconData info = _IconCode(0xe631);
+  static const IconData info = _IconData(0xe65a);
 
   /// 警告
-  static const IconData warning = _IconCode(0xe610);
+  static const IconData warning = _IconData(0xe65c);
 
   /// 笑脸
-  static const IconData smile = _IconCode(0xe62b);
+  static const IconData smile = _IconData(0xe65e);
+
+  /// 空数据
+  static const IconData empty = _IconData(0xe621);
+
+  /// 眼睛关闭
+  static const IconData eyeClose = _IconData(0xe65b);
+
+  /// 眼睛打开
+  static const IconData eyeOpen = _IconData(0xe659);
 }
 
-class _IconCode extends IconData {
-  const _IconCode(int codePoint)
-      : super(codePoint,
-            fontFamily: 'Icons',
-            matchTextDirection: true,
-            fontPackage: 'flutter_waya');
+class _IconData extends IconData {
+  const _IconData(super.codePoint)
+      : super(fontFamily: 'WayIcons', fontPackage: 'flutter_waya');
 }
 
 typedef Callback<T> = void Function();